backend/requirements.txt

requests==2.26.0
asgiref==3.5.2
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
djoser==2.0.5
Pillow==9.2.0
drf-extra-fields==3.4.0
django_cleanup==6.0.0
